+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 
 * Add the execution permission if possible
+* Show a specific message when user specify a script as game executable
 
 ### Changed